requirements.txt
accelerate
anthropic
attrdict
chess
gmpy2
google-generativeai
imageio
imageio-ffmpeg
lmdeploy
matplotlib
openai
pillow
pjtools
protobuf
pyqt5
pyqt5-tools
qwen_vl_utils
sentencepiece
timm
torch
torchvision
transformers
transformers_stream_generator
